+++ p/insa/index.js
@@ -25,7 +25,7 @@ const insaProfile = {
 
 	products: products,
 
-	journeyLeg: true,
+	trip: true,
 	radar: true
 }